feat: Add prerelease logic to build-and-publish-docker action
In the event a prerelease is being published, such as 1.2.3-alpha.4, it
will be marked with the `latest` tag as well, and the latest version of
`1` and `1.2`. This is undesirable since it's specifically marked as a
prerelease.

So, updated the logic to only tag the published image with the exact
prerelease version when it is in fact a prerelease. Any other semver
friendly tag like 1.2.3 will continue to be tagged the same way as
before.

This gives us peace of mind that version tags can be trusted as stable
tags to use even when we publish prereleases.