feat: Added multi-platform support to build-and-publish-docker
Added a `platforms` parameter, so we can use this convenient action with
the flexibility of handling different architectures.

Also, the `platforms` parameter defaults to `linux/amd64,linux/arm64`,
which is Intel/AMD and Apple Silicon, instead of just `linux/amd64`. So,
this makes it more convenient to automatically support Macs.

If an image can't be published for linux/arm64, or additional platforms
are desired, the `platforms` parameter can be passed to specify.

This improvement doesn't just involve passing the `platforms` parameter.
We also use the `docker/setup-buildx-action` so the proper engine will
handle the building, and `docker/setup-qemu-action` so that proper
emulation can be performed for the other platforms.

Ref: https://docs.docker.com/build/ci/github-actions/multi-platform/