feat: Add "Check Version Match" workflow as a version tagging safeguard
This workflow expects to be ran when pushing a tagged version, and it
will fail if the tagged version does not match the version in the
package.json file.

It supports a preceding "v" in the version, as long as the tag and the
package.json file are consistent, and it supports partial versions. All
of these examples would be valid if they are the same between the tag
and the package.json version:

1, 1.2, 1.2.3, 1.2.3-rc.1, v1, v1.2, v1.2.3, v1.2.3-rc.1.